+++ db/initiative_events_db.js
@@ -32,6 +32,7 @@ function parseContent(type, data) {
 		case "parliament-letter":
 		case "parliament-decision":
 		case "parliament-interpellation":
+		case "parliament-national-matter":
 		case "parliament-board-meeting":
 		case "parliament-committee-meeting": return JSON.parse(data)
 		case "text": return data
@@ -47,6 +48,7 @@ function serializeContent(type, data) {
 		case "parliament-letter":
 		case "parliament-decision":
 		case "parliament-interpellation":
+		case "parliament-national-matter":
 		case "parliament-board-meeting":
 		case "parliament-committee-meeting": return JSON.stringify(data)
 		case "text": return data