+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Oct 6, 2019
+- Adds support for matters of significant national importance events.
+
 ## Jul 23, 2019.
 - Adds [Riigikogu API](https://github.com/riigikogu-kantselei/api) syncing to fetch parliament proceedings and documents.
 - Imports external initiatives and their texts from the Riigikogu API.